linear/requirements.txt
# Linear MCP Server Requirements

# Linear API
linear-python==0.2.2

# Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Client
aiohttp>=3.8.4

# LangChain Components
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2

# OpenAI API
openai>=1.0.0

# Utilities
python-multipart>=0.0.6  # For FastAPI file uploads